+++ junos/resource_group_dual_system_test.go
@@ -33,8 +33,8 @@ func TestAccJunosGroupDualSystem_basic(t *testing.T) {
 							"system.0.backup_router_destination.#", "2"),
 						resource.TestCheckResourceAttr("junos_group_dual_system.testacc_node0",
 							"system.#", "1"),
-						resource.TestCheckResourceAttr("junos_group_dual_system.testacc_node0",
-							"system.0.backup_router_destination.1", "192.0.2.0/26"),
+						resource.TestCheckTypeSetElemAttr("junos_group_dual_system.testacc_node0",
+							"system.0.backup_router_destination.*", "192.0.2.0/26"),
 					),
 				},
 				{